Add support for tooltips to polygon overlays
Add a 'tooltip' attribute to GoogleMapPolygon, which can contain a simple
string to be used as a tooltip for the polygon.
If this attribute is set, add mouse over/move/out handlers to the Polygon,
and display the tooltip text in an appropriately-positioned div while the
mouse pointer is above the polygon.